requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
pymysql==1.0.2
python-dotenv==1.0.0
pyjwt==2.7.0
Flask-Migrate==4.1.0
alembic==1.16.4
Mako==1.3.10
reportlab==4.0.4
Pillow==10.0.1